+++ adapters/wagmi-connector/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @blocto/wagmi-connector
 
+## 1.1.1
+
+### Patch Changes
+
+- fec7693: Seperate EVM provider read-only methods
+- dca87b8: Fix 'connect' event handler payload type
+- b9f1eca: Handle EVM switch chain if not connected.
+- ac63830: wagmi three-shake config and rainbowkit use wagmi-connector
+- 2cacc63: Fix evm switch chain is connected logic
+- Updated dependencies [abd96b2]
+- Updated dependencies [fec7693]
+- Updated dependencies [dca87b8]
+- Updated dependencies [b9f1eca]
+- Updated dependencies [2cacc63]
+  - @blocto/sdk@0.5.5
+
+## 1.1.1-beta.2
+
+### Patch Changes
+
+- Fix evm switch chain is connected logic
+- Updated dependencies
+  - @blocto/sdk@0.5.5-beta.2
+
+## 1.1.1-beta.1
+
+### Patch Changes
+
+- Handle EVM switch chain if not connected.
+- Updated dependencies
+  - @blocto/sdk@0.5.5-beta.1
+
+## 1.1.1-beta.0
+
+### Patch Changes
+
+- fec7693: Seperate EVM provider read-only methods
+- Updated dependencies [fec7693]
+  - @blocto/sdk@0.5.5-beta.0
+
 ## 1.1.0
 
 ### Minor Changes